+++ internal/scheduling/nova/filter_weigher_pipeline_controller_test.go
@@ -11,10 +11,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/testutil"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
@@ -25,10 +22,7 @@ import (
 	api "github.com/cobaltcore-dev/cortex/api/external/nova"
 	"github.com/cobaltcore-dev/cortex/api/scheduling"
 	"github.com/cobaltcore-dev/cortex/api/v1alpha1"
-	"github.com/cobaltcore-dev/cortex/internal/knowledge/extractor/plugins/compute"
 	"github.com/cobaltcore-dev/cortex/internal/scheduling/lib"
-	"github.com/cobaltcore-dev/cortex/internal/scheduling/nova/crs"
-	hv1 "github.com/cobaltcore-dev/openstack-hypervisor-operator/api/v1"
 )
 
 // mockCandidateGatherer implements CandidateGatherer for testing
@@ -930,417 +924,6 @@ func TestFilterWeigherPipelineController_IgnorePreselection(t *testing.T) {
 			// Verify result is set when no error
 			if !tt.expectError && decision.Status.Result == nil {
 				t.Error("Expected result to be set but was nil")
-			}
-		})
-	}
-}
-
-func TestPickReservationSlot(t *testing.T) {
-	// vmMemBytes for a 4096 MiB flavor.
-	const vmMemBytes = int64(4096) * 1024 * 1024
-
-	makeSlot := func(name string, totalMemMiB, totalCPU, usedMemMiB, usedCPU int64) v1alpha1.Reservation {
-		var allocs map[string]v1alpha1.CommittedResourceAllocation
-		if usedMemMiB > 0 || usedCPU > 0 {
-			allocs = map[string]v1alpha1.CommittedResourceAllocation{
-				"vm-existing": {
-					Resources: map[hv1.ResourceName]resource.Quantity{
-						hv1.ResourceMemory: *resource.NewQuantity(usedMemMiB*1024*1024, resource.BinarySI),
-						hv1.ResourceCPU:    *resource.NewQuantity(usedCPU, resource.DecimalSI),
-					},
-				},
-			}
-		}
-		return v1alpha1.Reservation{
-			ObjectMeta: metav1.ObjectMeta{Name: name},
-			Spec: v1alpha1.ReservationSpec{
-				Resources: map[hv1.ResourceName]resource.Quantity{
-					hv1.ResourceMemory: *resource.NewQuantity(totalMemMiB*1024*1024, resource.BinarySI),
-					hv1.ResourceCPU:    *resource.NewQuantity(totalCPU, resource.DecimalSI),
-				},
-				CommittedResourceReservation: &v1alpha1.CommittedResourceReservationSpec{
-					Allocations: allocs,
-				},
-			},
-		}
-	}
-
-	tests := []struct {
-		name       string
-		candidates []v1alpha1.Reservation
-		want       string
-	}{
-		{
-			name:       "no candidates",
-			candidates: nil,
-			want:       "",
-		},
-		{
-			name:       "single slot fits fully",
-			candidates: []v1alpha1.Reservation{makeSlot("a", 8192, 8, 0, 0)},
-			want:       "a",
-		},
-		{
-			name:       "slot with zero remaining memory excluded",
-			candidates: []v1alpha1.Reservation{makeSlot("a", 8192, 8, 8192, 0)},
-			want:       "",
-		},
-		{
-			name: "picks slot with least remaining memory when both cover fully",
-			candidates: []v1alpha1.Reservation{
-				makeSlot("large", 8192, 8, 0, 0), // 8192 MiB remaining
-				makeSlot("small", 6144, 8, 0, 0), // 6144 MiB remaining
-			},
-			want: "small",
-		},
-		{
-			name: "name tiebreak when coverage and remaining memory are equal",
-			candidates: []v1alpha1.Reservation{
-				makeSlot("slot-b", 6144, 4, 0, 0),
-				makeSlot("slot-a", 6144, 4, 0, 0),
-			},
-			want: "slot-a",
-		},
-		{
-			name: "missing resource keys treated as zero remaining",
-			candidates: []v1alpha1.Reservation{
-				{
-					ObjectMeta: metav1.ObjectMeta{Name: "empty-res"},
-					Spec: v1alpha1.ReservationSpec{
-						CommittedResourceReservation: &v1alpha1.CommittedResourceReservationSpec{},
-					},
-				},
-			},
-			want: "",
-		},
-		{
-			name:       "partially used slot still usable",
-			candidates: []v1alpha1.Reservation{makeSlot("partial", 8192, 8, 2048, 2)}, // 6144 MiB remaining
-			want:       "partial",
-		},
-		{
-			name:       "overfill: slot smaller than VM is usable",
-			candidates: []v1alpha1.Reservation{makeSlot("a", 4096, 8, 2048, 0)}, // 2048 MiB remaining < vmMemBytes
-			want:       "a",
-		},
-		{
-			name: "overfill: full coverage preferred over partial",
-			candidates: []v1alpha1.Reservation{
-				makeSlot("partial", 4096, 8, 2048, 0), // 2048 MiB remaining, coverage=2048
-				makeSlot("full", 6144, 8, 0, 0),       // 6144 MiB remaining, coverage=4096 (full)
-			},
-			want: "full",
-		},
-		{
-			name: "overfill: highest partial coverage preferred",
-			candidates: []v1alpha1.Reservation{
-				makeSlot("low", 4096, 8, 2048, 0),  // 2048 MiB remaining, coverage=2048
-				makeSlot("high", 4096, 8, 1024, 0), // 3072 MiB remaining, coverage=3072
-			},
-			want: "high",
-		},
-		{
-			name:       "CPU exhausted slot is still usable (overfill applies to CPU too)",
-			candidates: []v1alpha1.Reservation{makeSlot("cpu-full", 8192, 2, 0, 2)}, // remCPU=0 but remMem>0
-			want:       "cpu-full",
-		},
-		{
-			name: "name tiebreak when memory equal (CPU no longer a tiebreak criterion)",
-			candidates: []v1alpha1.Reservation{
-				makeSlot("more-cpu", 6144, 8, 0, 0),
-				makeSlot("less-cpu", 6144, 4, 0, 0),
-			},
-			want: "less-cpu", // wins by name, not by CPU
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := crs.PickSlot(tt.candidates, vmMemBytes)
-			if got != tt.want {
-				t.Errorf("crs.PickSlot() = %q, want %q", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestRecordCRAllocation(t *testing.T) {
-	scheme := runtime.NewScheme()
-	if err := v1alpha1.AddToScheme(scheme); err != nil {
-		t.Fatalf("AddToScheme v1alpha1: %v", err)
-	}
-	if err := hv1.AddToScheme(scheme); err != nil {
-		t.Fatalf("AddToScheme hv1: %v", err)
-	}
-
-	const (
-		instanceUUID = "vm-uuid-1"
-		projectID    = "project-1"
-		flavorName   = "m1.large"
-		flavorGroup  = "m1"
-		selectedHost = "compute-1"
-	)
-
-	ratio := uint64(2048)
-	fg := compute.FlavorGroupFeature{
-		Name:           flavorGroup,
-		Flavors:        []compute.FlavorInGroup{{Name: flavorName, VCPUs: 2, MemoryMB: 4096}},
-		LargestFlavor:  compute.FlavorInGroup{Name: flavorName, VCPUs: 2, MemoryMB: 4096},
-		SmallestFlavor: compute.FlavorInGroup{Name: flavorName, VCPUs: 2, MemoryMB: 4096},
-		RamCoreRatio:   &ratio,
-	}
-
-	flavorKnowledge := func() *v1alpha1.Knowledge {
-		raw, err := v1alpha1.BoxFeatureList([]compute.FlavorGroupFeature{fg})
-		if err != nil {
-			t.Fatalf("BoxFeatureList: %v", err)
-		}
-		return &v1alpha1.Knowledge{
-			ObjectMeta: metav1.ObjectMeta{Name: "flavor-groups"},
-			Status: v1alpha1.KnowledgeStatus{
-				Raw: raw,
-				Conditions: []metav1.Condition{{
-					Type:               v1alpha1.KnowledgeConditionReady,
-					Status:             metav1.ConditionTrue,
-					Reason:             "Ready",
-					LastTransitionTime: metav1.Now(),
-				}},
-			},
-		}
-	}
-
-	makeReservation := func(name string, memMiB, cpus int64, proj, group, host string, allocs map[string]v1alpha1.CommittedResourceAllocation) *v1alpha1.Reservation {
-		return &v1alpha1.Reservation{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: name,
-				Labels: map[string]string{
-					v1alpha1.LabelReservationType: v1alpha1.ReservationTypeLabelCommittedResource,
-				},
-			},
-			Spec: v1alpha1.ReservationSpec{
-				Type:       v1alpha1.ReservationTypeCommittedResource,
-				TargetHost: host,
-				Resources: map[hv1.ResourceName]resource.Quantity{
-					hv1.ResourceMemory: *resource.NewQuantity(memMiB*1024*1024, resource.BinarySI),
-					hv1.ResourceCPU:    *resource.NewQuantity(cpus, resource.DecimalSI),
-				},
-				CommittedResourceReservation: &v1alpha1.CommittedResourceReservationSpec{
-					ProjectID:     proj,
-					ResourceGroup: group,
-					Allocations:   allocs,
-				},
-			},
-		}
-	}
-
-	makeRequest := func(uuid, proj, flavor string) api.ExternalSchedulerRequest {
-		return api.ExternalSchedulerRequest{
-			Spec: api.NovaObject[api.NovaSpec]{
-				Data: api.NovaSpec{
-					InstanceUUID: uuid,
-					Flavor: api.NovaObject[api.NovaFlavor]{
-						Data: api.NovaFlavor{Name: flavor, MemoryMB: 4096, VCPUs: 2},
-					},
-				},
-			},
-			Context: api.NovaRequestContext{ProjectID: proj},
-		}
-	}
-
-	makeDecision := func(host string, candidates ...string) *v1alpha1.Decision {
-		h := host
-		return &v1alpha1.Decision{
-			Status: v1alpha1.DecisionStatus{
-				Result: &v1alpha1.DecisionResult{
-					TargetHost:   &h,
-					OrderedHosts: candidates,
-				},
-			},
-		}
-	}
-
-	// makeCRObject creates a CommittedResource for the test project+flavorGroup with remaining capacity.
-	makeCRObject := func() *v1alpha1.CommittedResource {
-		return &v1alpha1.CommittedResource{
-			ObjectMeta: metav1.ObjectMeta{Name: "cr-1"},
-			Spec: v1alpha1.CommittedResourceSpec{
-				ProjectID:       projectID,
-				FlavorGroupName: flavorGroup,
-				State:           v1alpha1.CommitmentStatusConfirmed,
-				Amount:          *resource.NewQuantity(int64(8192)*1024*1024, resource.BinarySI),
-			},
-		}
-	}
-
-	// setReady marks a reservation as Ready so BuildCRSlotEvaluator indexes it.
-	setReady := func(res *v1alpha1.Reservation) *v1alpha1.Reservation {
-		res.Status.Conditions = []metav1.Condition{{
-			Type:               v1alpha1.ReservationConditionReady,
-			Status:             metav1.ConditionTrue,
-			Reason:             "Ready",
-			LastTransitionTime: metav1.Now(),
-		}}
-		return res
-	}
-
-	vmAlloc := func() map[string]v1alpha1.CommittedResourceAllocation {
-		return map[string]v1alpha1.CommittedResourceAllocation{
-			instanceUUID: {
-				CreationTimestamp: metav1.Now(),
-				Resources: map[hv1.ResourceName]resource.Quantity{
-					hv1.ResourceMemory: *resource.NewQuantity(int64(4096)*1024*1024, resource.BinarySI),
-					hv1.ResourceCPU:    *resource.NewQuantity(2, resource.DecimalSI),
-				},
-			},
-		}
-	}
-
-	tests := []struct {
-		name             string
-		objects          []client.Object
-		request          api.ExternalSchedulerRequest
-		decision         *v1alpha1.Decision
-		checkSlot        string
-		expectAllocation bool
-		expectedCRSlot   string // if non-empty, asserts PlacementCounter cr_slot label
-	}{
-		{
-			name: "writes allocation into matching reservation",
-			objects: []client.Object{
-				flavorKnowledge(),
-				makeCRObject(),
-				setReady(makeReservation("slot-1", 8192, 8, projectID, flavorGroup, selectedHost, nil)),
-			},
-			request:          makeRequest(instanceUUID, projectID, flavorName),
-			decision:         makeDecision(selectedHost, selectedHost),
-			checkSlot:        "slot-1",
-			expectAllocation: true,
-		},
-		{
-			name: "idempotent: UUID already in allocations",
-			objects: []client.Object{
-				flavorKnowledge(),
-				makeCRObject(),
-				setReady(makeReservation("slot-1", 8192, 8, projectID, flavorGroup, selectedHost, vmAlloc())),
-			},
-			request:          makeRequest(instanceUUID, projectID, flavorName),
-			decision:         makeDecision(selectedHost, selectedHost),
-			checkSlot:        "slot-1",
-			expectAllocation: true,
-		},
-		{
-			name: "PAYG: flavor not in any group",
-			objects: []client.Object{
-				flavorKnowledge(),
-				setReady(makeReservation("slot-1", 8192, 8, projectID, flavorGroup, selectedHost, nil)),
-			},
-			request:          makeRequest(instanceUUID, projectID, "unknown-flavor"),
-			decision:         makeDecision(selectedHost, selectedHost),
-			checkSlot:        "slot-1",
-			expectAllocation: false,
-		},
-		{
-			name: "no matching reservation: host mismatch",
-			objects: []client.Object{
-				flavorKnowledge(),
-				makeCRObject(),
-				setReady(makeReservation("slot-1", 8192, 8, projectID, flavorGroup, "other-host", nil)),
-			},
-			request:          makeRequest(instanceUUID, projectID, flavorName),
-			decision:         makeDecision(selectedHost, selectedHost),
-			checkSlot:        "slot-1",
-			expectAllocation: false,
-		},
-		{
-			name: "no slot fits: all capacity used",
-			objects: []client.Object{
-				flavorKnowledge(),
-				makeCRObject(),
-				setReady(makeReservation("slot-full", 4096, 2, projectID, flavorGroup, selectedHost,
-					map[string]v1alpha1.CommittedResourceAllocation{
-						"other-vm": {
-							Resources: map[hv1.ResourceName]resource.Quantity{
-								hv1.ResourceMemory: *resource.NewQuantity(int64(4096)*1024*1024, resource.BinarySI),
-								hv1.ResourceCPU:    *resource.NewQuantity(2, resource.DecimalSI),
-							},
-						},
-					})),
-			},
-			request:          makeRequest(instanceUUID, projectID, flavorName),
-			decision:         makeDecision(selectedHost, selectedHost),
-			checkSlot:        "slot-full",
-			expectAllocation: false,
-		},
-		{
-			name: "inactive CR (pending state) is filtered by IsActive(): no allocation",
-			objects: []client.Object{
-				flavorKnowledge(),
-				func() *v1alpha1.CommittedResource {
-					cr := makeCRObject()
-					cr.Spec.State = v1alpha1.CommitmentStatusPending
-					return cr
-				}(),
-				setReady(makeReservation("slot-1", 8192, 8, projectID, flavorGroup, selectedHost, nil)),
-			},
-			request:          makeRequest(instanceUUID, projectID, flavorName),
-			decision:         makeDecision(selectedHost, selectedHost),
-			checkSlot:        "slot-1",
-			expectAllocation: false,
-		},
-		{
-			name: "no knowledge CRD: logs error, no allocation",
-			objects: []client.Object{
-				makeReservation("slot-1", 8192, 8, projectID, flavorGroup, selectedHost, nil),
-			},
-			request:          makeRequest(instanceUUID, projectID, flavorName),
-			decision:         makeDecision(selectedHost, selectedHost),
-			checkSlot:        "slot-1",
-			expectAllocation: false,
-			expectedCRSlot:   "error",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fakeClient := fake.NewClientBuilder().
-				WithScheme(scheme).
-				WithObjects(tt.objects...).
-				Build()
-
-			counter := crs.NewPlacementCounter()
-			reg := prometheus.NewRegistry()
-			reg.MustRegister(counter)
-
-			controller := &FilterWeigherPipelineController{
-				BasePipelineController: lib.BasePipelineController[lib.FilterWeigherPipeline[api.ExternalSchedulerRequest]]{
-					Client: fakeClient,
-				},
-				CRRecorder: crs.Recorder{
-					Client:           fakeClient,
-					PlacementCounter: counter,
-				},
-			}
-
-			controller.CRRecorder.RecordPlacement(context.Background(), tt.decision, tt.request)
-
-			var res v1alpha1.Reservation
-			if err := fakeClient.Get(context.Background(), client.ObjectKey{Name: tt.checkSlot}, &res); err != nil {
-				t.Fatalf("Get reservation %q: %v", tt.checkSlot, err)
-			}
-			_, hasAlloc := res.Spec.CommittedResourceReservation.Allocations[instanceUUID]
-			if tt.expectAllocation && !hasAlloc {
-				t.Errorf("expected allocation for UUID %q but none found", instanceUUID)
-			}
-			if !tt.expectAllocation && hasAlloc {
-				t.Errorf("expected no allocation for UUID %q but one was written", instanceUUID)
-			}
-			if tt.expectedCRSlot != "" {
-				intent := string(tt.decision.Spec.Intent)
-				got := testutil.ToFloat64(counter.WithLabelValues("unknown", intent, tt.expectedCRSlot))
-				if got != 1 {
-					t.Errorf("PlacementCounter[flavor_group=unknown, intent=%q, cr_slot=%q] = %.0f, want 1",
-						intent, tt.expectedCRSlot, got)
-				}
 			}
 		})
 	}
